server/requirements.txt
appnope==0.1.0
decorator==4.0.9
Django==1.9.5
django-rest-assured==0.2.0
django-shell-plus==1.1.7
djangorestframework==3.3.3
elasticsearch==1.9.0
entrypoints==0.2
factory-boy==2.6.1
fake-factory==0.5.7
gnureadline==6.3.3
Jinja2==2.8
jsonschema==2.5.1
lxml==3.6.0
MarkupSafe==0.23
mistune==0.7.2
psycopg2==2.6.1
ptyprocess==0.5.1
python-dateutil==2.5.2
simplegeneric==0.8.1
six==1.10.0
terminado==0.6
traitlets==4.2.1
urllib3==1.15.1
xmlutils==1.3
-e git+https://github.com/18F/django-pg-fts.git#egg=django-pg-fts
